Merges episode 02 notebook content into static Markdown
The episode is static content, and includes a slideshow. Hence, merged
the content from the Jupyter Notebook into the Markdown-formatted episode.

The episode also uses a slideshow encoded as source in a Jupyter Notebook.
Added rules and defintions to Makefile, added RevealJS, for generating HTML
for the slides that will work from the server rather than only locally.
Revised README and instructor notes accordingly,